requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Evaluation dependencies
rouge-score==0.1.2
nltk==3.8.1
plotly==5.15.0
